Enable clippy::ptr_as_ptr and remove possible mutability changes
Since Rust 1.51 clippy::ptr_as_ptr is finally avaiblable (though not
enabled by default) to disallow raw pointer casts with `as`: those can
change mutability or convert to/from a non-pointer type. `.cast()`
exists for this reason, allowing the pointed-to type to change without
changing the fact that this is a raw pointer nor its mutability.